manual: extract getServerSideProps into a file
